rename components and add more style
